+++ epicore_utils/modules/visualize_protein.py
@@ -7,11 +7,12 @@ import matplotlib.ticker as tck
 from matplotlib.ticker import MaxNLocator
 import pandas as pd 
 import numpy as np
+import re
 import logging
 logger = logging.getLogger(__name__)
 
 
-def plot_peptide_length_dist(first_df: pd.DataFrame, second_df: pd.DataFrame, first_explode: str, second_explode: str, first_column: str, second_column: str, first_label: str, second_label: str) -> tuple[plt.figure,int,int]:
+def plot_peptide_length_dist(first_df: pd.DataFrame, second_df: pd.DataFrame, first_explode: str, second_explode: str, first_column: str, second_column: str, first_label: str, second_label: str, mod_pattern: str) -> tuple[plt.figure,int,int]:
     """Visualize the distribution of lengths of sequences.
     
     Args:
@@ -27,6 +28,9 @@ def plot_peptide_length_dist(first_df: pd.DataFrame, second_df: pd.DataFrame, fi
             second_df, for which the length distribution will be plotted.  
         first_label: Label for the values of first_column in the plot.
         second_label: Label for the values of second_column in the plot. 
+        mod_pattern: A comma separated string with delimiters for peptide
+            modifications
+
 
     Returns:
         A tuple including a matplotlib figure and two integers. The figure is a 
@@ -44,7 +48,16 @@ def plot_peptide_length_dist(first_df: pd.DataFrame, second_df: pd.DataFrame, fi
     seq_first = first_long[first_column]
     seq_second = second_long[second_column]
 
-    first_len = seq_first.map(lambda pep: len(pep)).to_list()
+    # remove modifications before accessing the sequence length
+    pattern = r'\(.*?\)'
+    peptide_seqs = seq_first.apply(lambda seq: re.sub(pattern,"",seq))
+    pattern = r'\[.*?\]'
+    peptide_seqs = peptide_seqs.apply(lambda seq: re.sub(pattern,"",seq))
+    if mod_pattern:
+        pattern = re.escape(mod_pattern.split(',')[0]) + r'.*?' + re.escape(mod_pattern.split(',')[1])
+        peptide_seqs = peptide_seqs.apply(lambda seq: re.sub(pattern,"",seq))
+
+    first_len = peptide_seqs.map(lambda pep: len(pep)).to_list()
     second_len = seq_second.map(lambda pep: len(pep)).to_list()
     
     ax.hist(first_len, bins=np.arange(5,50,1), color='grey', label=first_label, alpha=0.6)
@@ -78,7 +91,7 @@ def plot_protein_landscape(protein_df: pd.DataFrame, accession: str, proteome_di
 
     prot_landscape = [0 for _ in prot_seq]
     
-    fig_width = max(1,round(len(prot_landscape)/50))
+    fig_width = max(15,round(len(prot_landscape)/50))
     fig_height = 3
 
     max_intens = 0
@@ -97,18 +110,18 @@ def plot_protein_landscape(protein_df: pd.DataFrame, accession: str, proteome_di
         group_start = min(prot_row['grouped_peptides_start'].iloc[0][group])
         for idx, position in enumerate(landscape):
             ax.bar(group_start+int(idx),position,width=1, alpha=0.4, color=color)
-        for pos in range(prot_row['core_epitopes_start'].iloc[0][group],prot_row['core_epitopes_end'].iloc[0][group]):
+        for pos in range(prot_row['core_epitopes_start'].iloc[0][group],prot_row['core_epitopes_end'].iloc[0][group]+1):
             ax.bar(pos,0.5,width=1,color=color)
 
         max_intens = max(max_intens, max(landscape))
 
     ybins=max_intens/10
     ax.yaxis.set_major_locator(MaxNLocator(nbins=max(ybins, 5)))
-    ax.xaxis.set_major_locator(MaxNLocator(nbins=max(fig_width/10, 15)))
+    ax.xaxis.set_major_locator(MaxNLocator(nbins=max(fig_width/10, 25)))
 
-    ax.set_title('Number of peptides mapped to each amino acid position and core epitopes of protein {}'.format(accession))
+    ax.set_title('Landscape and consensus epitopes of the protein {}'.format(accession))
     ax.set_xlabel('Position in protein {}'.format(accession))
-    ax.set_ylabel('Number of mapped peptides')
+    ax.set_ylabel('Number of aligned peptides')
     return fig
 
 
@@ -126,9 +139,9 @@ def plot_core_mapping_peptides_hist(epitope_df: pd.DataFrame) ->  plt.figure:
     """
     fig, ax = plt.subplots(layout='constrained')
     n_peps = epitope_df['grouped_peptides_sequence'].apply(lambda sequences: len(sequences.split(',')))
-    ax.hist(n_peps,bins=np.arange(1,max(n_peps)+1,1))
+    ax.hist(n_peps,bins=np.arange(1,max(n_peps)+2,1))
     ax.set_yscale('log')
-    ax.set_xlabel('number of peptides contributing to epitope')
+    ax.set_xlabel('number of peptides contributing to each consensus epitope')
     ax.set_ylabel('count')
     return fig